Add files via upload
The growth of supermarkets in most populated cities are increasing and market competitions are also high. The dataset is one of the historical sales of supermarket company which has recorded in 3 different branches for 3 months data. This project aims to explore the Supermarket Sales data to understand top performing branches and products, sales trend of different products, customer behaviour. The aims are to study how sales strategies can be improved and optimized. The dataset was obtained from the Kaggle Supermarket Sales.